// web/src/app/api/generated/src/models/ParticipantMeasurementPresetView.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Motion Ecosystem API
 * API for participant and specialist workflows
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface ParticipantMeasurementPresetView
 */
export interface ParticipantMeasurementPresetView {
  /**
   *
   * @type {Array<string>}
   * @memberof ParticipantMeasurementPresetView
   */
  allowedUnits?: Array<string>;
  /**
   *
   * @type {string}
   * @memberof ParticipantMeasurementPresetView
   */
  defaultUnit?: string;
  /**
   *
   * @type {string}
   * @memberof ParticipantMeasurementPresetView
   */
  id?: string;
  /**
   *
   * @type {string}
   * @memberof ParticipantMeasurementPresetView
   */
  label?: string;
  /**
   *
   * @type {Array<string>}
   * @memberof ParticipantMeasurementPresetView
   */
  requiredContextFields?: Array<string>;
}

/**
 * Check if a given object implements the ParticipantMeasurementPresetView interface.
 */
export function instanceOfParticipantMeasurementPresetView(
  value: object,
): value is ParticipantMeasurementPresetView {
  return true;
}

export function ParticipantMeasurementPresetViewFromJSON(
  json: any,
): ParticipantMeasurementPresetView {
  return ParticipantMeasurementPresetViewFromJSONTyped(json, false);
}

export function ParticipantMeasurementPresetViewFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): ParticipantMeasurementPresetView {
  if (json == null) {
    return json;
  }
  return {
    allowedUnits: json['allowedUnits'] == null ? undefined : json['allowedUnits'],
    defaultUnit: json['defaultUnit'] == null ? undefined : json['defaultUnit'],
    id: json['id'] == null ? undefined : json['id'],
    label: json['label'] == null ? undefined : json['label'],
    requiredContextFields:
      json['requiredContextFields'] == null ? undefined : json['requiredContextFields'],
  };
}

export function ParticipantMeasurementPresetViewToJSON(
  json: any,
): ParticipantMeasurementPresetView {
  return ParticipantMeasurementPresetViewToJSONTyped(json, false);
}

export function ParticipantMeasurementPresetViewToJSONTyped(
  value?: ParticipantMeasurementPresetView | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    allowedUnits: value['allowedUnits'],
    defaultUnit: value['defaultUnit'],
    id: value['id'],
    label: value['label'],
    requiredContextFields: value['requiredContextFields'],
  };
}